cns windows tests
